+++ data/trait.go
@@ -32,7 +32,7 @@ var Traits []Trait = []Trait{
 	{Code: 'F', Name: "Fast Growth", DoubleName: "Rapid Growth", Ingredients: []Ingredient{
 		{"drinkCanMegaCrush", 2},
 	}},
-	{Code: 'E', Name: "Explosive", DoubleName: "Extremely Explosive", IncompatibleTraits: []rune{'B', 'R'}, Ingredients: []Ingredient{
+	{Code: 'E', Name: "Explosive", DoubleName: "Extremely Explosive", IncompatibleTraits: []rune{'B', 'R', 'T'}, Ingredients: []Ingredient{
 		{"resourceScrapIron", 4},
 		{"resourceGunPowder", 4},
 		{"resourceDuctTape", 1},
@@ -40,7 +40,7 @@ var Traits []Trait = []Trait{
 	{Code: 'R', Name: "Renewable", IncompatibleTraits: []rune{'R', 'E'}, Ingredients: []Ingredient{
 		{"drinkJarPureMineralWater", 10},
 	}},
-	{Code: 'T', Name: "Thorny", DoubleName: "Extra Thorny", Ingredients: []Ingredient{
+	{Code: 'T', Name: "Thorny", DoubleName: "Extra Thorny", IncompatibleTraits: []rune{'E'}, Ingredients: []Ingredient{
 		{"resourceScrapIron", 10},
 		{"resourceNail", 10},
 	}},
@@ -82,7 +82,7 @@ func (t *Trait) GetTraitDescription() string {
 		return fmt.Sprintf(`%s: further doubles crop yield.`,
 			t.Name)
 	case 'U':
-		return fmt.Sprintf(`%s: fused with mushroom dna, alowing growth without the need for sunlight.`,
+		return fmt.Sprintf(`%s: fused with mushroom dna, allowing growth without the need for sunlight.`,
 			t.Name)
 	case 'F':
 		return fmt.Sprintf(`%s: reaches maturity in half the time.`,