+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - set hotbox repair resources
 - resolve issue with explosive trait collision
 - update readme
+- make Thorny and Explosive traits incompatible
 
 ## [0.6.0] - 2022-05-19